README.md*
This website was created for a multifunctional private medical clinic called "Plus Med" which offers various medical services, to fulfill the day-to-day virtual necessities of the clinic. The website consists of six main pages that give all the essential information about the clinic. On the Header part of the website pages like "About us", "Doctors", "Service" "Contact", and "Reservation" are located, with each page giving specific information related to that section. To the most left header of the website, on the page "About us" information about why people specifically choose this particular clinic, also the rating of the clinic, and its achievements. . On the page "Service"  the information related to the medical equipment types used in the clinic to treat patients can be found. Moreover, the page about the "Service" website allows users to get acquainted with general information related to the clinic, and get primary information about leading doctors of the clinic, their areas of specialties, years of experience, and universities they graduated. It also gives an advantage for online reservations in advance. The next page is called "Contact" where contact information
 of the clinic can be found for reservation purposes. Finally, the last page of the website is made for "Reservation" which is used to book doctor meetings, different room reservations and reservations for different treatment procedures.